+++ src/models/schema/identitySchema.ts
@@ -1,0 +1,11 @@
+import mongoose from 'mongoose'
+
+export const IdentitySchema = new mongoose.Schema(
+    {
+        NationtalIndentityNumber: {
+            type: Number,
+        },
+       
+    },
+    { _id: false }
+)
